// frontend/src/backend.d.ts
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
export type Time = bigint;
export interface DailyMediaEntry {
    id: string;
    createdAt: Time;
    user: Principal;
    relatedDay: Time;
    mediaType: MediaType;
}
export interface JournalEntry {
    id: string;
    createdAt: Time;
    user: Principal;
    audioUrl: string;
    transcription: string;
    notes: string;
    dayType?: DayType;
    photos: Array<ExternalBlob>;
}
export interface Fichier {
    id: bigint;
    nom: string;
    url: string;
    contenu: ExternalBlob;
    description: string;
    taille: bigint;
    typeMime: string;
    proprietaire: Principal;
    dateAjout: Time;
}
export interface TimeEntryInput {
    id: string;
    heuresTrajet: bigint;
    date: Time;
    description: string;
    startMorning: bigint;
    endAstreinte?: bigint;
    interventionSlots: Array<InterventionSlot>;
    startAstreinte?: bigint;
    endMorning: bigint;
    endAfternoon: bigint;
    typeOfDay: DayType;
    heuresRepas: bigint;
    startAfternoon: bigint;
}
export interface Totals {
    heuresTrajet: bigint;
    heuresTravailNormales: bigint;
    heuresAstreinte: bigint;
    heuresRepas: bigint;
}
export interface InterventionSlot {
    endHour: bigint;
    endMinute: bigint;
    startMinute: bigint;
    startHour: bigint;
}
export interface PdfReportData {
    titre: string;
    enteteTableau: Array<string>;
    periode: string;
    totaux: {
        heuresTrajet: string;
        heuresTravailNormales: string;
        heuresAstreinte: string;
        heuresRepas: string;
    };
    exportTimestamp: Time;
    lignesTableau: Array<Array<string>>;
}
export type MediaType = {
    __kind__: "audio";
    audio: ExternalBlob;
} | {
    __kind__: "photo";
    photo: ExternalBlob;
};
export interface UserProfile {
    name: string;
    email: string;
}
export interface TimeEntry {
    id: string;
    heuresTrajet: bigint;
    date: Time;
    user: Principal;
    description: string;
    startMorning: bigint;
    endAstreinte?: bigint;
    interventionSlots: Array<InterventionSlot>;
    startAstreinte?: bigint;
    endMorning: bigint;
    endAfternoon: bigint;
    typeOfDay: DayType;
    heuresRepas: bigint;
    startAfternoon: bigint;
}
export enum DayType {
    conge = "conge",
    work = "work",
    astreinte = "astreinte"
}